After limited environment testing, it is tailored to our needs, which is to detect the presence of music in the ambient audio captured by the live microphone. The root mean square (RMS) analysis originally used in the V3-1 code is not a suitable method for detecting the presence of music.
This update changes the analysis method from root mean square (RMS) analysis to short-time Fourier transform (STFT). Analyze the frequency components of the audio signal by converting the time domain audio data to the frequency domain to determine whether there are prominent frequency peaks. The final implementation determines whether there is a melody pattern or musical content in the ambient audio captured by the microphone. If it exists, it proves that there is at least audio content that can be considered a melody in the environmental audio. At this time, we operate obs to switch audio sources.

Advantages: It uses a relatively simple algorithm to analyze music frequencies. While completing basic analysis, it reduces the consumption of live broadcast machine functions.
Disadvantages: The analysis method used is not robust enough, and it is difficult to detect some long melodies. And under the interference of complex external environmental noise, it is impossible to ensure that accurate frequency analysis can be completed.
At present, the recording time of a single analysis unit is 6s. If the length of a single chord exceeds 3s, the detection cannot be ensured (the principle is to detect whether there is a repeated spectrum. To detect it, it must be repeated at least once.). The length of the recording time of this analysis unit is directly proportional to the system response speed. That is, when set to 6s, it will be judged every 6s whether to switch the audio source. Continuing to lengthen the recording time will cause the system to wait too long for each response, making it meaningless.