This update introduces the python audio analysis library librosa. Replace the relevant code sections that perform the melody analysis using NumPy with the librosa.feature.chroma_stft function. This function computes the chromagram from a waveform or power spectrogram. By analyzing the chromagram, we can detect the presence of harmonic content, which can indicate the presence of music or melody patterns in the audio signal.
Advantages: It uses librosa, a special library developed for music analysis, to analyze music frequencies. The algorithm is more targeted and accurate. The ability to resist external noise interference has been improved.
Disadvantages: The computational performance consumption of the live broadcast machine increased by the algorithm. Some longer melodies are difficult to detect.
The recording time of a single analysis unit is 6 seconds. If a single chord is longer than 6s, detection is not guaranteed. The length of the analysis unit recording time affects the system response speed. After each recording is completed, determine whether to switch the audio source and delete the audio recorded in the previous section. Extending the recording time will lengthen the system response interval, making it impossible to switch in time to avoid copyright issues.